requirements.txt
# Streamlit
streamlit>=1.28.0

# Database
snowflake-connector-python>=3.0.0
snowflake-sqlalchemy>=1.5.0

# Google AI
google-genai==0.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0

# Date/Time
pytz>=2023.3
